Let max_msg_per_second be ignored by default
Summary:
change default value of max_msg_per_second to be 0 and ignore
it if it is so. Fallout is if someone is unknowingly relying on the default value of 100000 for throttling they will have to tune max_queue_size. But this will serve them well going forward.

Reviewed By: agiardullo

Test Plan:
run testsuite.php. The change is simple enough to not require
testing.